Relevant BUGIDs: task 15788
Purpose of commit: more autoconf stuff

Commit summary:
---------------
Basic support for autoconf in all three libraries and one module:
pam_permit.

I've moved the pam_misc.h header sources to make them more consistent
with libpam and libpamc. Later, when we tackle the top level makefile
we'll need to fix the way we build the include stuff.

At this point, one should 'autoconf' in the top level directory and
then cd to each of the libraries to build them. The rest of the package
is not expected to build yet.