Relevant BUGIDs: task 15788
Purpose of commit: new features

Commit summary:
---------------
This commit should complete the first attempt at autoconfing
the modules. I've consolodated all the install_conf scripts
into one and thus am removing a number of files.
As of this commit, each of the libraries and all of the modules
should build and install and remove as per the directions from
autoconf. (that is toplevel ./configure .)