Relevant BUGIDs: 117476, see task 15788
Purpose of commit: bugfix

Commit summary:
---------------
See bug id, but pelov rightly points out that the build tree does not
provide the location of the pam_misc header when compiling the libpamc
library.

I belive the correct fix is to not include pam_misc.h, but to include
_pam_macros.h instead. The correct -I path is provided in at least
the autoconf CVS branch, which I intend to merge with the main trunk
before releasing 0.73.