Relevant BUGIDs: 117474, task 15788
Purpose of commit: bugfix

Commit summary:
---------------
As pelov rightly points out in the bug report, there was a problem
with the -I include for upperLOWER.c (or lack there of). I've
implemented the fix he suggests and also removed the gcc warning about
the archaic use of three arguments to main() in this filter.