Relevant BUGIDs: 126036
Purpose of commit: cleanup

Commit summary:
---------------
Changes format of pam_unix 'authentication failure' log messages for
purposes of consistency and clarity.

Note that we do not log the 'user=' item in the case of an invalid username
unless the audit flag is turned on.

Note also that the way we currently log these failures isn't i18n-friendly.
I imagine this will need correcting at some point in the future.