Relevant BUGIDs: 493294
Purpose of commit: bugfix and document

Commit summary:
---------------
The login limit counting seemed to have its math incorrect.
The default has been changed to the sane one, and a new module argument
has been added "utmp_early" that returns the module to its old behavior.

Thanks to Berend De Schouwe for getting to the bottom of this issue.