Relevant BUGIDs: 517064
Purpose of commit: feature

Commit summary:
---------------
document old feature and add '\]' parsing to make it a better feature.
The feature is that we can accept spaces in module arguments by enclosing
the whole argument inside square brackets. For example a module argument
like this:

  "[hello [you\], this is me]"

will be parsed as

  "hello [you], this is me"

Not very interesting, but you get the idea.
Thanks to Russell Kliese for requesting this.