Relevant BUGIDs: mail
Purpose of commit: bugfix

Commit summary:
---------------

The following reported by mail was fixed:

"I've been in trouble using the '%group' feature in the limits.conf file.
There are priorities when settings the limits and the order is as
enumerated here. The problem is that the initial value is
LIMITS_DEF_NONE, so neither LIMITS_DEF_ALL nor LIMITS_DEF_ALLGROUP
can ever be set."