build: fix crypt configure with musl libc
This patch fixes a case where it finds crypt defined in libc (musl) but
not in specified libraries then it ends up assigning
LIBCRYPT="-l" which then goes into makefile cause all sort of problems
e.g.

ld: cannot find -l-m32
| collect2: error: ld returned 1 exit status
The reason is that -l appears on commandline with
out any library and compiler treats the next argument as library name
whatever it is.

Signed-off-by: Khem Raj <raj.khem@gmail.com>
Signed-off-by: Maxin B. John <maxin.john@intel.com>
Signed-off-by: James Hilliard <james.hilliard1@gmail.com>