Fix pam_xauth when using pam_mktemp.
For some reason when using su, getgid won't return the calling user's gid but 0 (root).
pam_mktemp will cause the xauthority file to be created on a folder where either uid 0 or the user's gid are required to access it.
As a result of those two conditions, xauth will be run with the user's uid but gid 0 and therefore fail acquiring the magic cookies locking the call to su for 30 seconds.

In this commit we replace all calls to getuid and getgid to accesses to the cached structure from pam_modutil_getpwuid which is used at the beggining.
This will make xauth run using both the user's uid and gid and fix this issue.

As a side effect, having less syscalls may speed up loging slightly (but to a nearly unnoticeable level).

For more information, see https://bugs.gentoo.org/show_bug.cgi?id=585550