pam_userdb: allow key_only with hashes, use DB API
This change allows using key_only with hashed passwords, which didn't
work before but was not documented.

I also migrated away from the historic dbm API, which means that the
db=/path/to/file parameter now has to be specified as a full filename,
including any suffix.

The crypt= parameter was also renamed to hash=