pam_env: ignore / skip lines beginning with '='
pam_putenv in libpam/pam_env.c makes a special exception for an '=' as
the first character, aborting with PAM_BAD_ITEM. When combined with the
truncation to 1023 characters, if the 1024th character is an '=', this
can cause unexpected system lockout. Note, this would also happen if the
line started with an '='.

Related to issue #31.