pam_unix: Better implementation of checksalt for expired passphrases.
libxcrypt can be build-time configured to support (or not support)
various hashing methods.  Future versions will also have support for
runtime configuration by the system's vendor and/or administrator.

For that reason users should be notified by pam if they cannot log
into their account anymore because of such a change in the system's
configuration of libxcrypt.

Also check for malformed des hashes like "$2...", which might have
been generated by unsafe base64 encoding functions as in glibc prior
to version 2.17 and get silently rejected by the crypt(3) function
of implemented in glibc 2.17 and later or libxcrypt version 4.1.0
and later.

The previous (and thus dropped) implementation of this feature had
some flaws, which are fixed in this implementation.

* modules/pam_unix/passverify.c (checksalt_authtok_expired): New func.
* modules/pam_unix/passverify.c (check_shadow_expiry): Add checksalt
for expired passphrases using the new function.