pam_securetty: silently ignore missing securetty file
Fedora stopped shipping /etc/securetty years ago and it is also not shipped in
Debian and Debian derivatives either because nowadays it is not found to be
useful.
Since the default state is not having /etc/securetty on the system PAM noticing
its absence also should not show up in syslog either.