pam_umask: build-time usergroups option default
This change adds a configure option to set the default value of the
usergroups option (of the pam_umask module) at build-time.

Distributions usually makes the decision if usergroups should be used or
not. This allows them to control the built-in default value, without
having to ship the value in a config file (cluttering up the view
of actually relevant user/system configuration overrides).