doc: fix the description of stack jump effects
Every stack jump, besides the jump itself, has a side effect which is
one of 'ignore', 'ok', or 'bad'.  Unfortunately, the side effect is far
from obvious because it depends on the PAM function call, and the
documentation that contradicts the implementation does not help either.

* doc/man/pam.conf-syntax.xml (actionN): Rewrite the description
of stack jump effects to match the implementation.

Fixes: 871a6e14d65c3c446ae0af51166dabc7a47a2b56