Drop Travis CI support
The last build of Linux-PAM by travis-ci.org was in March of 2020,
and travis-ci.org says that all building is ceased since June of 2021.

Given that in foreseeable future travis-ci.com is not going to welcome
free software projects, there is no use to keep Travis CI support
in the tree.

* .travis.yml: Remove.

Link: https://blog.travis-ci.com/2020-11-02-travis-ci-new-billing