pam_motd: do not rely on all filesystems providing a filetype
When using scandir() to look for MOTD files to display, we wrongly
relied on all filesystems providing a filetype.  This is a fix to divert
to lstat() when we have no filetype.  To maintain MT safety, it isn't
possible to use lstat() in the scandir() filter function, so all of the
filtering has been moved to an additional loop after scanning all the
motd dirs.
Also, remove superfluous alphasort from scandir(), we are doing
a qsort() later.

Resolves: https://github.com/linux-pam/linux-pam/issues/455