security: Add Landlock file restriction
This patch adds a minimalist Landlock support to PAM which restricts the
SUID unix_chkpwd binary to a readonly access to the whole filesystem.
The goal here is to provide security in depth limiting access to the few
resources this program needs limiting uncontrolled privileged access to
the system.

The patch should evolve to support finer restrictions on other PAM
modules/binaries.

* libpam/include/security/pam_sandbox.h: Add a minimalist Landlock file
  restriction function.

* modules/pam_unix/unix_chkpwd.c: Add a read-only access restriction
  to the filesystem using Landlock.

Signed-off-by: Nicolas Bouchinet <nicolas.bouchinet@ssi.gouv.fr>