pam_unix: use helper on skeleton root entry from libnss-systemd
libnss-systemd started[1] to synthesize a root entry with an invalid
password.  On Debian since version 251.3-2[2] systemd enables the
systemd module for shadow and gsshadow in nsswitch.conf.  This leads on
SELinux enabled systems, where the files module can not access the
shadow file, to the root account being locked.

Fall back to the unix_chkpwd helper in case the root user was queried
and the response was a locked password ("!*"[3]).

[1]: https://github.com/systemd/systemd/commit/f43a19ecd6e3415e
[2]: https://salsa.debian.org/systemd-team/systemd/-/commit/bf9a307cfcbe62ab4fbcf2198e6e628a1bca211b
[3]: https://github.com/systemd/systemd/blob/b622e95f2f59fcb58e23ddafed745eee26a0f52f/src/nss-systemd/nss-systemd.c#L34